--- app/layout.tsx
@@ -1,28 +1,0 @@
-import type React from "react"
-import "@/app/globals.css"
-import { Inter } from "next/font/google"
-
-import { ThemeProvider } from "@/components/theme-provider"
-
-const inter = Inter({ subsets: ["latin"] })
-
-export const metadata = {
-  title: "Nimbus - Better Cloud Storage",
-  description: "A modern cloud storage solution",
-}
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={inter.className}>
-        <ThemeProvider defaultTheme="dark" attribute="class">
-          {children}
-        </ThemeProvider>
-      </body>
-    </html>
-  )
-}
