+++ apps/ui/app/page.tsx
@@ -1,0 +1,22 @@
+import { FileBrowser } from "@/ui/components/file-browser"
+import { Header } from "@/ui/components/header"
+import { Sidebar } from "@/ui/components/sidebar"
+import { UploadButton } from "@/ui/components/upload-button"
+
+export default function DrivePage() {
+  return (
+    <div className="min-h-screen bg-background text-foreground flex flex-col">
+      <Header />
+      <div className="flex flex-1 overflow-hidden">
+        <Sidebar />
+        <main className="flex-1 overflow-auto p-6">
+          <div className="flex items-center justify-between mb-6">
+            <h1 className="text-2xl font-semibold">My Files</h1>
+            <UploadButton />
+          </div>
+          <FileBrowser />
+        </main>
+      </div>
+    </div>
+  )
+}
