+++ apps/ui/lib/auth.ts
@@ -1,0 +1,9 @@
+import { betterAuth } from "better-auth";
+import { drizzleAdapter } from "better-auth/adapters/drizzle";
+import { db } from "@/server/src/db/index"; // your drizzle instance
+ 
+export const auth = betterAuth({
+    database: drizzleAdapter(db, {
+        provider: "pg", // or "mysql", "sqlite"
+    })
+});
