+++ package.json
@@ -1,34 +1,29 @@
 {
-  "name": "0-drive",
+  "name": "nimbus",
   "version": "0.1.0",
   "private": true,
+  "workspaces": [
+    "apps/*",
+    "packages/*"
+  ],
   "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@radix-ui/react-avatar": "^1.1.10",
-    "@radix-ui/react-collapsible": "^1.1.11",
-    "@radix-ui/react-dialog": "^1.1.14",
-    "@radix-ui/react-dropdown-menu": "^2.1.15",
-    "@radix-ui/react-progress": "^1.1.7",
-    "@radix-ui/react-separator": "^1.1.7",
-    "@radix-ui/react-slot": "^1.2.3",
-    "@radix-ui/react-tabs": "^1.1.12",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "lucide-react": "^0.511.0",
-    "next": "15.2.4",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "tailwind-merge": "^3.3.0"
+    "dev:ui": "bun --cwd=apps/ui run dev",
+    "dev:server": "bun --cwd=apps/server run dev",
+    "dev": "bun run dev:ui & bun run dev:server",
+    "build:ui": "bun --cwd=apps/ui run build",
+    "build:server": "bun --cwd=apps/server run build",
+    "build": "bun run build:ui && bun run build:server",
+    "start:ui": "bun --cwd=apps/ui run start",
+    "start:server": "bun --cwd=apps/server run start",
+    "start": "bun run start:ui & bun run start:server",
+    "lint": "bun run lint:ui & bun run lint:server",
+    "lint:ui": "bun --cwd=apps/ui run lint",
+    "lint:server": "bun --cwd=apps/server run lint"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
+    "@types/bun": "latest",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
